# trace:  Span in noop.Start is no longer allocated