@tttoad 2024-05-31
```
func BenchmarkNoopInstance(b *testing.B) {
	tracer := NewTracerProvider().Tracer("")
	spanCtx := trace.SpanContext{}

	ctx := trace.ContextWithSpanContext(context.Background(), spanCtx)

	for i := 0; i < b.N; i++ {
		for j := 0; j < 10; j++ {
			newCtx, span := tracer.Start(ctx, "")
			ctx = newCtx
			span.End()
		}
	}
}
// non-single instance
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/trace/noop
BenchmarkNoopInstance-10         1146366               966.8 ns/op          1280 B/op         20 allocs/op
PASS
ok      go.opentelemetry.io/otel/trace/noop     2.351s

// isingle instance
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/trace/noop
BenchmarkNoopInstance-10         2038297               555.7 ns/op           480 B/op         10 allocs/op
PASS
ok      go.opentelemetry.io/otel/trace/noop     1.890s
```