@pellared 2024-05-31
Looks like https://github.com/open-telemetry/opentelemetry-go/pull/5049

Can you add the benchmark?