@pellared 2024-06-03
Can you please add a changelog entry similar to https://github.com/open-telemetry/opentelemetry-go/pull/5049 and update the PR description with benchmark results?