@MrAlias 2024-06-03
Please provide the [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) report in the PR description for the included benchmark.