@tttoad 2024-06-04
@MrAlias 
<img width="937" alt="image" src="https://github.com/open-telemetry/opentelemetry-go/assets/14021602/34fecd23-fcfa-4dc9-806c-3c6c6aa89ed8">

I changed the first description of the PR.
Maybe I'm misunderstanding you?