@tttoad 2024-06-04
@MrAlias Sorry, first time using `benchstat`. I misunderstood.